Resolve circular dependency issue
