+++ app/react/package.json
@@ -57,6 +57,7 @@
     "webpack": "^4.33.0"
   },
   "devDependencies": {
+    "@storybook/client-api": "6.0.0-alpha.15",
     "@types/mini-css-extract-plugin": "^0.9.0",
     "@types/node": "^13.5.3",
     "@types/webpack": "^4.41.0"
