+++ lib/client-api/src/types.ts
@@ -67,3 +67,13 @@ export interface GetStorybookKind {
   fileName: string;
   stories: GetStorybookStory[];
 }
+
+// This really belongs in lib/core, but that depends on lib/ui which (dev) depends on app/react
+// which needs this type. So we put it here to avoid the circular dependency problem.
+export type RenderContext = StoreItem & {
+  forceRender: boolean;
+
+  showMain: () => void;
+  showError: (error: { title: string; description: string }) => void;
+  showException: (err: Error) => void;
+};
